[legacy] refactor initialize
