+++ colossalai/context/__init__.py
@@ -1,2 +1,8 @@
 from .config import Config, ConfigException
-from .moe_context import MOE_CONTEXT
+
+# from .moe_context import MOE_CONTEXT
+
+__all__ = [
+    'Config',
+    'ConfigException',
+]
