+++ colossalai/utils/cuda.py
@@ -1,7 +1,10 @@
 #!/usr/bin/env python
 # -*- encoding: utf-8 -*-
 
+from typing import Optional
+
 import torch
+import torch.distributed as dist
 
 
 def set_to_cuda(models):
@@ -23,7 +26,7 @@ def set_to_cuda(models):
 def get_current_device() -> torch.device:
     """
     Returns currently selected device (gpu/cpu).
-    If cuda available, return gpu, otherwise return cpu.    
+    If cuda available, return gpu, otherwise return cpu.
     """
     if torch.cuda.is_available():
         return torch.device(f'cuda:{torch.cuda.current_device()}')
@@ -45,3 +48,9 @@ def empty_cache():
     """
     if torch.cuda.is_available():
         torch.cuda.empty_cache()
+
+
+def set_device(index: Optional[int] = None) -> None:
+    if index is None:
+        index = dist.get_rank() % torch.cuda.device_count()
+    torch.cuda.set_device(index)
