+++ colossalai/zero/gemini/memory_tracer/memory_monitor.py
@@ -5,7 +5,6 @@ from time import sleep, time
 
 import torch
 
-from colossalai.legacy.utils import colo_device_memory_used
 from colossalai.utils import get_current_device
 
 
@@ -111,6 +110,7 @@ class AsyncMemoryMonitor(MemoryMonitor):
         return max_usage
 
     def _measure_usage(self):
+        from colossalai.legacy.utils import colo_device_memory_used
         max_usage = 0
         while self.keep_measuring:
             max_usage = max(
