+++ tests/test_legacy/test_layers/test_2d/test_2d.py
@@ -18,8 +18,8 @@ from checks_2d.check_layer_2d import (
 )
 from checks_2d.check_operation_2d import check_AB, check_ABT, check_ATB
 
-from colossalai.initialize import launch
 from colossalai.legacy.core import global_context as gpc
+from colossalai.legacy.initialize import launch
 from colossalai.logging import disable_existing_loggers
 from colossalai.testing import rerun_if_address_is_in_use, spawn
