+++ tests/test_legacy/test_pipeline/test_middleware_1f1b.py
@@ -7,9 +7,9 @@ import torch.distributed.rpc as rpc
 from rpc_test_utils import DAG_MLP, MLP
 from torch._C._distributed_rpc import _is_current_rpc_agent_set
 
-from colossalai import launch
 from colossalai.fx import ColoTracer
 from colossalai.fx.passes.adding_split_node_pass import balanced_split_pass, split_with_split_nodes_pass
+from colossalai.legacy import launch
 from colossalai.legacy.pipeline.middleware.adaptor import get_fx_topology
 from colossalai.legacy.pipeline.pipeline_process_group import ppg
 from colossalai.legacy.pipeline.rpc._pipeline_schedule import OneFOneBPipelineEngine
