+++ tests/test_legacy/test_trainer/test_pipeline/test_pipeline_schedule.py
@@ -11,9 +11,9 @@ from torchvision.datasets import CIFAR10
 from torchvision.models import resnet18
 
 import colossalai
-from colossalai.initialize import launch
 from colossalai.legacy.context import ParallelMode
 from colossalai.legacy.core import global_context as gpc
+from colossalai.legacy.initialize import launch
 from colossalai.legacy.utils import get_dataloader, print_rank_0
 from colossalai.testing import rerun_if_address_is_in_use, spawn
 
@@ -63,7 +63,7 @@ def run_schedule(rank, world_size, port):
     optimizer = torch.optim.Adam(model.parameters(), lr=0.001, weight_decay=0)
 
     # initialize
-    engine, train_dataloader, _, _ = colossalai.initialize(model, optimizer, criterion, train_dataloader)
+    engine, train_dataloader, _, _ = colossalai.legacy.initialize(model, optimizer, criterion, train_dataloader)
 
     # build pipeline schedule
     schedule = engine.schedule
